mv pre/post process
